경매이미지기능 미완성

+++ backend/src/main/java/com/ilgoojo/backend/repository/AuctionRepository.java
@@ -3,6 +3,6 @@ package com.ilgoojo.backend.repository;
 import com.ilgoojo.backend.entity.Auction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionRepository extends JpaRepository<Auction, Long> {
+public interface AuctionRepository extends JpaRepository<Auction, Integer> {
 
 }
